XW-3180 | recirculation clean up, phantom update to work with new node version
