+++ extensions/wikia/Recirculation/Recirculation.i18n.php
@@ -9,7 +9,6 @@ $messages['en'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'Wiki',
 	'recirculation-fandom-title' => 'Trending Fandom Articles',
 	'recirculation-fandom-subtitle' => 'Also Recommended',
-	'recirculation-incontent-title' => 'Top Links',
 	'recirculation-discussion-title' => 'Top Discussions',
 	'recirculation-discussion-link-text' => 'View All Discussions',
 	'recirculation-discussions-new' => 'Start a New Discussion',
@@ -26,7 +25,6 @@ $messages['qqq'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'Text to display as a tag on top of Wiki content',
 	'recirculation-fandom-title' => 'Title to display at the top of the Fandom recirculation module',
 	'recirculation-fandom-subtitle' => 'Title to display on some design variations of the Fandom recirculation module',
-	'recirculation-incontent-title' => 'Title to display beside the incontent widget',
 	'recirculation-discussion-title' => 'Title to display at the top of the Discussion recirculation module',
 	'recirculation-discussion-link-text' => 'Text for the link to other discussions',
 	'recirculation-discussions-new' => 'Text to include on a CTA to start a new discussion',
@@ -43,7 +41,6 @@ $messages['de'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'Wiki',
 	'recirculation-fandom-title' => 'Fandom-Artikel im Trend',
 	'recirculation-fandom-subtitle' => 'Ebenfalls empfohlen',
-	'recirculation-incontent-title' => 'Top-Links',
 	'recirculation-discussion-title' => 'Top-Diskussionen',
 	'recirculation-discussion-link-text' => 'Alle Diskussionen anzeigen',
 	'recirculation-discussions-new' => 'Eine neue Diskussionen beginnen',
@@ -60,7 +57,6 @@ $messages['es'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'Wiki',
 	'recirculation-fandom-title' => 'Artículos de Fandom en Tendencia',
 	'recirculation-fandom-subtitle' => 'También Recomendado',
-	'recirculation-incontent-title' => 'Enlaces Top',
 	'recirculation-discussion-title' => 'Discusiones Top',
 	'recirculation-discussion-link-text' => 'Ver todas las Discusiones',
 	'recirculation-discussions-new' => 'Comenzar una nueva Discusión',
@@ -77,7 +73,6 @@ $messages['fr'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'Wikia',
 	'recirculation-fandom-title' => 'Articles de fandom tendance',
 	'recirculation-fandom-subtitle' => 'Également recommandé',
-	'recirculation-incontent-title' => 'Liens les plus populaires',
 	'recirculation-discussion-title' => 'Discussions les plus populaires',
 	'recirculation-discussion-link-text' => 'Voir toutes les discussions',
 	'recirculation-discussions-new' => 'Lancer une discussion',
@@ -94,7 +89,6 @@ $messages['it'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'Wiki',
 	'recirculation-fandom-title' => 'Articoli Trend Fandom',
 	'recirculation-fandom-subtitle' => 'Consigliato anche',
-	'recirculation-incontent-title' => 'I principali link',
 	'recirculation-discussion-title' => 'Le principali discussioni',
 	'recirculation-discussion-link-text' => 'Visualizza tutte le discussioni',
 	'recirculation-discussions-new' => 'Avvia una nuova discussione',
@@ -111,7 +105,6 @@ $messages['ja'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'ウィキ',
 	'recirculation-fandom-title' => '人気のファンダム記事',
 	'recirculation-fandom-subtitle' => 'こちらもおすすめ',
-	'recirculation-incontent-title' => 'トップリンク',
 	'recirculation-discussion-title' => 'おすすめのトーク',
 	'recirculation-discussion-link-text' => 'トークをすべて表示',
 	'recirculation-discussions-new' => '新しいトークを開始',
@@ -128,7 +121,6 @@ $messages['nl'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'Wiki',
 	'recirculation-fandom-title' => 'Trending Fandom Articles',
 	'recirculation-fandom-subtitle' => 'Also Recommended',
-	'recirculation-incontent-title' => 'Top Links',
 	'recirculation-discussion-title' => 'Top Discussions',
 	'recirculation-discussion-link-text' => 'View All Discussions',
 	'recirculation-discussions-new' => 'Start a New Discussion',
@@ -145,7 +137,6 @@ $messages['pl'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'Wiki',
 	'recirculation-fandom-title' => 'Najpopularniejsze artykuły ze strony Fandom',
 	'recirculation-fandom-subtitle' => 'Również polecane',
-	'recirculation-incontent-title' => 'Polecane linki',
 	'recirculation-discussion-title' => 'Najpopularniejsze dyskusje',
 	'recirculation-discussion-link-text' => 'Pokaż wszystkie dyskusje',
 	'recirculation-discussions-new' => 'Rozpocznij nową dyskusję',
@@ -162,7 +153,6 @@ $messages['pt'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'Wiki',
 	'recirculation-fandom-title' => 'Artigos do Fandom em alta',
 	'recirculation-fandom-subtitle' => 'Também recomendado',
-	'recirculation-incontent-title' => 'Links populares',
 	'recirculation-discussion-title' => 'Discussões populares',
 	'recirculation-discussion-link-text' => 'Veja todas as Discussões',
 	'recirculation-discussions-new' => 'Comece uma nova Discussão',
@@ -179,7 +169,6 @@ $messages['ru'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'Вики',
 	'recirculation-fandom-title' => 'Популярные статьи с Fandom',
 	'recirculation-fandom-subtitle' => 'Также рекомендуем',
-	'recirculation-incontent-title' => 'Близкие по теме статьи',
 	'recirculation-discussion-title' => 'Популярные обсуждения',
 	'recirculation-discussion-link-text' => 'Все обсуждения',
 	'recirculation-discussions-new' => 'Начать новое обсуждение',
@@ -196,7 +185,6 @@ $messages['zh-hans'] = array(
 	'recirculation-impact-footer-wiki-tag' => '维基',
 	'recirculation-fandom-title' => '粉丝流行原创文章',
 	'recirculation-fandom-subtitle' => '推荐',
-	'recirculation-incontent-title' => '热门推荐',
 	'recirculation-discussion-title' => '热门讨论',
 	'recirculation-discussion-link-text' => '展示所有讨论',
 	'recirculation-discussions-new' => '发起一个新讨论',
@@ -213,7 +201,6 @@ $messages['zh-hant'] = array(
 	'recirculation-impact-footer-wiki-tag' => 'Wiki',
 	'recirculation-fandom-title' => '粉絲流行原創文章',
 	'recirculation-fandom-subtitle' => '推薦',
-	'recirculation-incontent-title' => '精選',
 	'recirculation-discussion-title' => '熱門討論',
 	'recirculation-discussion-link-text' => '顯示所有討論',
 	'recirculation-discussions-new' => '開始一個新討論',
