--- extensions/wikia/Recirculation/styles/rail.scss
@@ -1,36 +1,0 @@
-@import 'skins/shared/color';
-@import 'skins/shared/mixins/transform';
-
-.WikiaRail .recirculation-rail {
-	.recirculation-sponsored-footer {
-		font-size: 12px !important;
-	}
-
-	.item {
-		margin-bottom: 15px;
-	}
-
-	.image-thumbnail {
-		margin-bottom: 7px;
-		max-height: 166px;
-		max-width: 268px;
-		overflow: hidden;
-
-	}
-
-	.title {
-		font-size: 15px;
-	}
-
-	.metadata {
-		display: none;
-	}
-
-	a {
-		color: $color-text;
-
-		&:hover {
-			color: $color-links;
-		}
-	}
-}
