+++ extensions/wikia/Recirculation/styles/recirculation.scss
@@ -1,9 +1,6 @@
 @import 'horizontal';
-@import 'rail';
 @import 'premium-rail';
 @import 'discussions';
-@import 'perfect-scrollbar';
-@import 'scroller';
 @import 'impact-footer';
 
 @import 'skins/shared/color';
