--- extensions/wikia/Recirculation/styles/scroller.scss
@@ -1,79 +1,0 @@
-@import 'skins/shared/color';
-@import 'skins/shared/mixins/chevron';
-@import 'skins/shared/mixins/transition';
-@import 'skins/shared/mixins/transform';
-
-$recirculation-scroller-color: #ccc;
-
-.recirculation-horizontal.recirculation-scroller {
-	padding: 40px 0 25px;
-
-	.item {
-		margin-right: 20px;
-
-		&.is-video .thumbnail::after {
-			height: 36px;
-			width: 36px;
-		}
-	}
-
-	&:hover {
-		.scroller-arrow {
-			opacity: .95;
-		}
-	}
-}
-
-.recirculation-horizontal .ps-container > .ps-scrollbar-x-rail {
-	bottom: -4px;
-	opacity: 1;
-
-	.ps-scrollbar-x {
-		background-color: $recirculation-scroller-color;
-	}
-}
-
-$recirculation-scroller-chevron-size: 30px;
-$recirculation-scroller-arrow-offset: 10px;
-$recirculation-scroller-shadow: rgba(0, 0, 0, .5);
-
-.scroller-arrow {
-	@include transition-shorthand(opacity .3s ease-in-out);
-
-	background: $color-page;
-	border-radius: 2px;
-	box-shadow: -2px 0 4px $recirculation-scroller-shadow;
-	cursor: pointer;
-	display: block;
-	opacity: 0;
-	padding: $recirculation-scroller-chevron-size $recirculation-scroller-chevron-size / 2;
-	position: absolute;
-	top: 55px;
-	z-index: 1;
-
-	.icon {
-		@include right-chevron($recirculation-scroller-chevron-size, 2px, right, $color-links);
-	}
-
-	&::after {
-		background-color: $color-page;
-		bottom: 0;
-		content: '';
-		display: block;
-		left: $recirculation-scroller-arrow-offset;
-		position: absolute;
-		right: -$recirculation-scroller-arrow-offset;
-		top: 0;
-		z-index: -1;
-	}
-}
-
-.scroller-next {
-	right: -$recirculation-scroller-arrow-offset;
-}
-
-.scroller-prev {
-	@include transform(scale(-1, 1));
-	left: -$recirculation-scroller-arrow-offset;
-	right: auto;
-}
