--- extensions/wikia/Recirculation/templates/Recirculation_index.mustache
@@ -1,15 +1,0 @@
-<section class="recirculation-rail module">
-	<h2>{{title}}</h2>
-	<ul class="thumbnails">
-		{{#posts}}
-			<li class="title-thumbnail">
-				<a href="{{url}}">
-					<div class="image image-thumbnail fluid small">
-						<img src="{{image_url}}" alt="{{title}}" itemprop="thumbnail">
-					</div>
-					<div class="title" title="{{title}}">{{title}}</div>
-				</a>
-			</li>
-		{{/posts}}
-	</ul>
-</section>
