--- extensions/wikia/Recirculation/templates/client/incontent.mustache
@@ -1,21 +1,0 @@
-<div class="recirculation-horizontal recirculation-incontent recirculation-unit">
-	<h3>{{title}}</h3>
-
-	<div class="items">
-		{{#items}}
-		<div
-		    data-flag="{{flag}}"
-		    class="item item-{{source}}{{#isVideo}} is-video{{/isVideo}}"
-		    data-index="{{index}}"
-		    data-source="{{source}}"
-		    {{#id}} data-id="{{id}}" {{/id}}
-		    {{#meta}} data-meta="{{meta}}" {{/meta}}
-		>
-			<a title="{{title}}" href="{{url}}">
-				<div class="thumbnail"><img src="{{thumbnail}}" alt="{{title}}"></div>
-				<h4>{{title}}</h4>
-			</a>
-		</div>
-		{{/items}}
-	</div>
-</div>
