--- extensions/wikia/Recirculation/templates/client/scroller.mustache
@@ -1,27 +1,0 @@
-<div class="recirculation-horizontal recirculation-incontent recirculation-scroller recirculation-unit">
-	<h3>{{title}}</h3>
-
-	<div class="items-container">
-		<div class="items">
-			{{#items}}
-			<div class="item item-{{source}}{{#isVideo}} is-video{{/isVideo}}"
-				data-index="{{index}}"
-				data-source="{{source}}"
-				{{#id}} data-id="{{id}}" {{/id}}
-				{{#meta}} data-meta="{{meta}}" {{/meta}}
-			>
-				<a title="{{title}}" href="{{url}}">
-					<div class="thumbnail placeholder-thumbnail">
-						{{#thumbnail}}
-							<img src="{{thumbnail}}" alt="{{title}}">
-						{{/thumbnail}}
-					</div>
-					<h4>{{title}}</h4>
-				</a>
-			</div>
-			{{/items}}
-		</div>
-		<div class="scroller-arrow scroller-prev" data-direction="prev"><div class="icon"></div></div>
-		<div class="scroller-arrow scroller-next" data-direction="next"><div class="icon"></div></div>
-	</div>
-</div>
