+++ package.json
@@ -27,9 +27,10 @@
     "karma-html2js-preprocessor": "~0.1.0",
     "karma-jasmine": "~0.3.2",
     "karma-junit-reporter": "~0.2.2",
-    "karma-phantomjs-launcher": "~0.1.4",
+    "karma-phantomjs-launcher": "1.0.0",
     "karma-requirejs": "~0.2.2",
     "karma-script-launcher": "~0.1.0",
+    "phantomjs-prebuilt": "2.1.8",
     "requirejs": "~2.1.15",
     "jscs-jsdoc": "~1.2.0"
   }
